+++ inc/lilyNotation/measureLayout.ts
@@ -1,10 +1,21 @@
 
+enum LayoutType {
+	Ordinary = "ordinary",
+	Full = "full",
+	Conservative = "conservative",
+	Once = "once",
+};
+
+
 interface MeasureLayout {
-	//serialize (type: string): number[];
+	serialize (type: LayoutType): number[];
 };
 
 
 type MeasureSeq = MeasureLayout[];
+
+
+const spreadMeasureSeq = (seq: MeasureSeq, type: LayoutType = LayoutType.Ordinary): number[] => [].concat(...seq.map(layout => layout.serialize(type)));
 
 
 class SingleMLayout implements MeasureLayout {
@@ -12,6 +23,11 @@ class SingleMLayout implements MeasureLayout {
 
 	constructor (measure: number) {
 		this.measure = measure;
+	}
+
+
+	serialize (): number[] {
+		return [this.measure];
 	}
 };
 
@@ -46,6 +62,11 @@ class BlockMLayout implements MeasureLayout {
 			}
 		}
 	}
+
+
+	serialize (type: LayoutType): number[] {
+		return spreadMeasureSeq(this. seq, type);
+	}
 };
 
 
@@ -53,17 +74,74 @@ class VoltaMLayout implements MeasureLayout {
 	times: number;
 	body: MeasureSeq = [];
 	alternates: MeasureSeq[];
+
+
+	serialize (type: LayoutType): number[] {
+		const bodySeq = spreadMeasureSeq(this.body);
+
+		if (this.alternates) {
+			const alternateSeqs = this.alternates.map(seq => spreadMeasureSeq(seq));
+			const lastAlternateSeq = alternateSeqs[alternateSeqs.length - 1];
+
+			switch (type) {
+			case LayoutType.Ordinary:
+				return bodySeq.concat(...alternateSeqs);
+
+			case LayoutType.Conservative:
+			case LayoutType.Full: {
+				const priorSeq = [].concat(...Array(this.times - 1).fill(null).map((_, i) => [
+					...bodySeq,
+					...alternateSeqs[i % (this.times - 1)],
+				]));
+
+				return [
+					...priorSeq,
+					...bodySeq,
+					...lastAlternateSeq,
+				];
+			}
+
+			case LayoutType.Once:
+				return [
+					...bodySeq,
+					...lastAlternateSeq,
+				];
+			}
+		}
+		else {
+			switch (type) {
+			case LayoutType.Ordinary:
+			case LayoutType.Conservative:
+			case LayoutType.Once:
+				return bodySeq;
+
+			case LayoutType.Full:
+				return [].concat(...Array(this.times).fill(null).map(() => bodySeq));
+			}
+		}
+
+		console.warn("the current case not handled:", type, this);
+	}
 };
 
 
 class ABAMLayout implements MeasureLayout {
 	main: MeasureLayout;
 	rest: MeasureSeq = [];
+
+
+	serialize (type: LayoutType): number[] {
+		void type;
+
+		// TODO:
+		return [];
+	}
 };
 
 
 
 export {
+	LayoutType,
 	MeasureLayout,
 	SingleMLayout,
 	BlockMLayout,