+++ changelog.md
@@ -2,8 +2,6 @@
 ## TODO
 	implicit notes matching with sheet document
 		match lilyNotation with articulated MIDI (ignored repeats), then unfold repeats on lilyNotation, include D.C., D.S.
-			parsing repeat mark among measure index
-				MusicBlock.measureLayout: parse ABA measure layout
 	staffNotation
 		remove note time from staff document
 	lilyInterpreter
@@ -24,6 +22,8 @@
 
 
 *	2020.9.6
+
+	+	measureLayout.ts: ABAMLayout.serialize implemented.
 
 	+	notation.ts: output option of measure layout type.